requirements.txt
requests~=2.27.1
websockets~=10.2
pandas~=1.4.2
numpy~=1.22.3
python-telegram-bot